requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch>=2.0.0
transformers>=4.34.0
pillow>=10.0.0
pydantic>=2.0.0